backend/requirements.txt
# TradeIQ Backend - Design Document v3.0
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
dj-database-url>=2.1
psycopg2-binary>=2.9
channels>=4.0
channels[daphne]>=4.0
atproto>=0.0.18
python-dotenv>=1.0
